+++ ShiftRpg/Effects/DamageEffect.cs
@@ -7,11 +7,11 @@ namespace ShiftRpg.Effects;
 public record DamageEffect(
     Team AppliesTo,
     SourceTag Source,
-    int Damage,
+    float Damage,
     ICollection<float> AdditiveIncreases,
     ICollection<float> MultiplicativeIncreases) : IEffect
 {
-    public DamageEffect(Team AppliesTo, SourceTag Source, int Damage)
+    public DamageEffect(Team AppliesTo, SourceTag Source, float Damage)
         : this(AppliesTo, Source, Damage, new List<float>(), new List<float>()) { }
 
     public Guid EffectId { get; } = Guid.NewGuid();
